feat(v4): client feedback batch 1 — quick wins (backend)
From the client's review PDF:
- users: sales reps were not creatable from the UI (backend already allowed it) and users had no
  edit at all — add PATCH /users/{id} (name/role/scope/active + optional password reset).
- governorates: could only be seeded — add POST/PATCH so new areas can be entered.
- points are now FRACTIONAL (client: "6 pieces = 1 point"): point_value/delta -> NUMERIC(18,3),
  point_service arithmetic moved to Decimal; the API returns points as decimal strings like money.
- item categories: Item.category + list filter, managed from Settings (new `item_category` lookup).
- contacts: extra phone numbers for customers AND suppliers (new contact_phone table);
  supplier.address; customer governorate_id/markaz/address (detailed address).
- rule: a plumber's responsible rep must be after-sales staff; bootstrap now seeds an
  `aftersales` user and the demo seed assigns plumbers to them.

Migration 0013 + startup _widen_columns() so live DBs get the fractional points type.
Tests: 6 new + full suite 301 passed.

+++ backend/src/api/points.py
@@ -1,6 +1,8 @@
 """Points router (T024): derived balance + conversion. FR-007."""
 from __future__ import annotations
 
+from decimal import Decimal
+
 from fastapi import APIRouter, Depends, HTTPException, status
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
@@ -16,7 +18,7 @@ router = APIRouter(tags=["points"], prefix="/customers")
 
 class PointBalanceOut(BaseModel):
     customer_id: int
-    balance: int
+    balance: Decimal
     derived: bool = True
